+++ app/Models/Instance/AuditLog.php
@@ -1,0 +1,80 @@
+<?php
+
+namespace App\Models\Instance;
+
+use App\Helpers\LogHelper;
+use App\Helpers\DateHelper;
+use App\Models\Account\Account;
+use App\Models\User\User;
+use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Eloquent\Relations\BelongsTo;
+
+class AuditLog extends Model
+{
+    protected $table = 'audit_logs';
+
+    /**
+     * The attributes that are mass assignable.
+     *
+     * @var array
+     */
+    protected $fillable = [
+        'account_id',
+        'author_id',
+        'author_name',
+        'action',
+        'objects',
+        'audited_at',
+    ];
+
+    /**
+     * The attributes that should be mutated to dates.
+     *
+     * @var array
+     */
+    protected $dates = [
+        'audited_at',
+    ];
+
+    /**
+     * Get the Account record associated with the audit log.
+     *
+     * @return BelongsTo
+     */
+    public function account()
+    {
+        return $this->belongsTo(Account::class);
+    }
+
+    /**
+     * Get the User record associated with the audit log.
+     *
+     * @return BelongsTo
+     */
+    public function author()
+    {
+        return $this->belongsTo(User::class);
+    }
+
+    /**
+     * Get the JSON object.
+     *
+     * @return array
+     * @param mixed $value
+     */
+    public function getObjectAttribute($value)
+    {
+        return json_decode($this->objects);
+    }
+
+    /**
+     * Get the content of the audit log, if defined.
+     *
+     * @return string
+     * @param mixed $value
+     */
+    public function getContentAttribute($value): string
+    {
+        return LogHelper::processAuditLog($this);
+    }
+}