+++ database/migrations/2020_02_03_015403_create_audit_log_table.php
@@ -1,0 +1,39 @@
+<?php
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+class CreateAuditLogTable extends Migration
+{
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        Schema::create('audit_logs', function (Blueprint $table) {
+            $table->bigIncrements('id');
+            $table->unsignedInteger('account_id');
+            $table->unsignedInteger('author_id')->nullable();
+            $table->string('author_name');
+            $table->string('action');
+            $table->text('objects');
+            $table->datetime('audited_at');
+            $table->timestamps();
+            $table->foreign('account_id')->references('id')->on('accounts')->onDelete('cascade');
+            $table->foreign('author_id')->references('id')->on('users')->onDelete('set null');
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        Schema::dropIfExists('audit_logs');
+    }
+}